+++ src/features/home/components/app-updater/app-updater.component.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect } from "react"
+import React, { useCallback, useEffect, useRef } from "react"
 import {
   withServiceWorkerUpdater,
   ServiceWorkerUpdaterProps,
@@ -8,16 +8,12 @@ import { StyledUpdateButton, StyledAppVersion } from "./app-updater.styles"
 
 const AppUpdater: React.FC<ServiceWorkerUpdaterProps> = (props) => {
   const { newServiceWorkerDetected, onLoadNewServiceWorkerAccept } = props
-  const [currentVersion, setCurrentVersion] = React.useState("2.0.0")
+  const appVersion = useRef("")
 
   useEffect(() => {
-    try {
-      const versionNumber = process?.env?.REACT_APP_VERSION
-      if (versionNumber) {
-        setCurrentVersion(`v${versionNumber}`)
-      }
-    } catch (error) {
-      console.error(error)
+    const versionNumber = process.env.REACT_APP_VERSION
+    if (versionNumber) {
+      appVersion.current = `v${versionNumber}`
     }
   }, [])
 
@@ -28,20 +24,17 @@ const AppUpdater: React.FC<ServiceWorkerUpdaterProps> = (props) => {
   return (
     <>
       {newServiceWorkerDetected ? (
-        <span>
-          New version detected {" → "}
-          <StyledUpdateButton
-            type="button"
-            onClick={onLoadNewServiceWorkerAccept}
-          >
-            Update
-          </StyledUpdateButton>
-        </span>
-      ) : (
+        <StyledUpdateButton
+          type="button"
+          onClick={onLoadNewServiceWorkerAccept}
+        >
+          Update
+        </StyledUpdateButton>
+      ) : appVersion.current ? (
         <StyledAppVersion type="button" onClick={handleRefresh}>
-          {currentVersion} ↻
+          {appVersion.current} ↻
         </StyledAppVersion>
-      )}
+      ) : null}
     </>
   )
 }