+++ src/features/home/components/pokoy/pokoy.component.tsx
@@ -1,18 +1,16 @@
-import { useState, useEffect, useCallback } from "react"
+import { useState, useCallback } from "react"
 import { User } from "@firebase/auth"
-import { LOCAL_CACHE_FIELD_NAME, SECS_IN_MIN } from "shared/constants"
-import { firestore } from "../../firebase-init"
+import { useAppDispatch, useAppSelector } from "store"
+import useNoSleep from "shared/hooks/use-no-sleep"
+import { mainScreenActions } from "../../store/main-screen.slice"
+import { setMeditationThunk } from "../../store/main-screen.thunks"
 import { TimerButton } from "../timer-button/timer-button.component"
 import { Countdown } from "../countdown/countdown.component"
+import { Sound } from "../sound.component"
 import { Tips } from "../tips"
-import { sendSessionFromLocalStore } from "./writeSessionToServer"
-import { PokoySession } from "shared/types"
-import { BottomTextWrapper, TopTextWrapper, Wrapper } from "./pokoy.styles"
-import { Sound } from "features/home/components/sound.component"
 import { FibSpiral } from "../fib-spiral/fib-spiral.component"
-import { setMeditationThunk } from "features/user-stats/user-stats.thunks"
-import { useAppDispatch } from "store"
-import useNoSleep from "shared/hooks/use-no-sleep"
+import { BottomTextWrapper, TopTextWrapper, Wrapper } from "./pokoy.styles"
+import { SECS_IN_MIN } from "shared/constants"
 
 interface Props {
   user: User
@@ -23,9 +21,10 @@ interface Props {
 export const Pokoy: React.FC<Props> = ({ user, authLoading }) => {
   const [currentTimerId, setCurrentTimerId] = useState<number | null>(null)
   const [timerDiff, setTimerDiff] = useState<number>(0)
-  // TODO: extract to store
-  const [isStarted, setStartedFlag] = useState(false)
-  useNoSleep(isStarted)
+
+  const isTimerStarted =
+    useAppSelector((state) => state.mainScreen.timerStatus) === "started"
+  useNoSleep(isTimerStarted)
   const dispatch = useAppDispatch()
 
   const minutes = Math.floor(timerDiff / SECS_IN_MIN)
@@ -36,7 +35,8 @@ export const Pokoy: React.FC<Props> = ({ user, authLoading }) => {
       if (!isCurrentTimerIdExist) throw Error("currentTimerId is not exist")
 
       window.clearInterval(currentTimerId)
-      setStartedFlag(false)
+
+      dispatch(mainScreenActions.setTimerStatus(false))
       setTimerDiff(0)
 
       const isSessionLongerThanMinute = timerDiff > SECS_IN_MIN
@@ -63,40 +63,24 @@ export const Pokoy: React.FC<Props> = ({ user, authLoading }) => {
 
   const startTimer = useCallback(() => {
     const startInSeconds = Math.round(Date.now() / 1000)
-    setStartedFlag(true)
+    dispatch(mainScreenActions.setTimerStatus(true))
 
     const newTimerId = window.setInterval(
       () => handleTimer(startInSeconds),
       100
     )
     setCurrentTimerId(newTimerId)
-  }, [handleTimer])
+  }, [dispatch, handleTimer])
 
   const handleClick = useCallback(() => {
     setTimerDiff(0)
 
-    if (isStarted) {
+    if (isTimerStarted) {
       return finishTimer(timerDiff)
     } else {
       return startTimer()
     }
-  }, [finishTimer, isStarted, startTimer, timerDiff])
-
-  // TODO: extract function in useEffect from component or extract custom hook
-  useEffect(() => {
-    const storedAfterFailurePokoySession = window?.localStorage.getItem(
-      LOCAL_CACHE_FIELD_NAME
-    )
-
-    if (storedAfterFailurePokoySession) {
-      const lastSession = JSON.parse(
-        storedAfterFailurePokoySession
-      ) as PokoySession
-
-      sendSessionFromLocalStore(firestore, user, lastSession)
-      window?.localStorage.removeItem(LOCAL_CACHE_FIELD_NAME)
-    }
-  }, [user])
+  }, [finishTimer, isTimerStarted, startTimer, timerDiff])
 
   return (
     <Wrapper>
@@ -106,7 +90,7 @@ export const Pokoy: React.FC<Props> = ({ user, authLoading }) => {
 
       <TimerButton
         handleTimerClick={handleClick}
-        isTimerStarted={isStarted}
+        isTimerStarted={isTimerStarted}
         authLoading={authLoading}
       >
         <Sound progress={timerDiff} />
@@ -114,7 +98,9 @@ export const Pokoy: React.FC<Props> = ({ user, authLoading }) => {
       </TimerButton>
 
       <BottomTextWrapper>
-        {!authLoading && <Tips minutes={minutes} isTimerStarted={isStarted} />}
+        {!authLoading && (
+          <Tips minutes={minutes} isTimerStarted={isTimerStarted} />
+        )}
       </BottomTextWrapper>
     </Wrapper>
   )