+++ src/features/home/components/sound.component.tsx
@@ -1,6 +1,5 @@
 import useSound from "use-sound"
 import React from "react"
-import gongSfx from "shared/assets/sounds/gong.mp3"
 import bellSfx from "shared/assets/sounds/meditation-bell-sound.mp3"
 import { FIB_NUMS_FOR_TIMER } from "shared/constants"
 
@@ -9,9 +8,6 @@ interface Props {
 }
 
 export const Sound: React.FC<Props> = ({ progress }) => {
-  const [playGong] = useSound(gongSfx, {
-    volume: 0.03,
-  })
   const [playBell] = useSound(bellSfx, {
     volume: 0.3,
   })
@@ -21,18 +17,13 @@ export const Sound: React.FC<Props> = ({ progress }) => {
     const minutes = progress / 60
     const isStart = minutes === 0
     const isFibNum = FIB_NUMS_FOR_TIMER.includes(minutes)
-    const isFinish = minutes === 21
 
     if (isStart) return
-
-    if (isFinish) {
-      return playGong()
-    }
 
     if (isFibNum) {
       return playBell()
     }
-  }, [progress, playGong, playBell])
+  }, [progress, playBell])
 
   return null
 }